finel tool calling with float

+++ tools.py
@@ -1,38 +1,38 @@
 class tools:
-    def add(self,a:int,b:int)->int:
+    def add(self,a:float,b:float)->float:
         """ Add a and b.
 
         Args :
-        a: first int
-        b: second int
+        a: first float
+        b: second float
         """
         return a+b
 
-    def subtract(self,a:int,b:int)->int:
+    def subtract(self,a:float,b:float)->float:
         """ subtract b from a
 
         Args :
-        a: first int
-        b: second int
+        a: first float
+        b: second float
         """
         return a-b
     
-    def multiply(self,a: int , b:int)-> int:
+    def multiply(self,a: float , b:float)-> float:
         """ Multiply a and b
 
         Args :
-        a: first int
-        b: second int
+        a: first float
+        b: second float
         """
         return a*b
 
-    def divide(self,a :int, b:int)->float:
+    def divide(self,a :float, b:float)->float:
 
         """ Divide a by b
 
         Args :
-        a: first int
-        b: second int
+        a: first float
+        b: second float
         """
         return a/b
